requirements.txt
aiobotocore==2.21.1
channels[daphne]==4.2.2
channels-redis==4.2.1
django==5.2
django-allauth[socialaccount]==65.7.0
django-constance==4.3.2
django-cors-headers==4.7.0
django-filter==25.1
django-oauth-toolkit==3.0.1
django-storages[boto3]==1.14.6
djangorestframework==3.16.0
freezegun==1.5.1
gunicorn==23.0.0
pillow==11.1.0
psycopg[pool]==3.2.6
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-django==4.11.1
python-slugify==8.0.4
sentry-sdk[django]==2.25.1
uvicorn==0.34.0
wait-for-it==2.3.0
scrapy==2.13.1
celery==5.5.2
scrapy-djangoitem==1.1.1
